Change start page to have panels
issue: MatterHackers/MCCentral#3516
Revise start page
