+++ PartPreviewWindow/StartPage/ExplorerBar.cs
